Module Fadbad

Interface to library FADBADml

Basic types

module type OpS = Fadbad__.Types.OpS

Type of a module of operators

module type OrderedOpS = Fadbad__.Types.OrderedOpS

Type of a module of operators over ordered elements

Basic operators

module OpFloat = Fadbad__.Op.Float

Operators over float elements

module OrderedFloat = Fadbad__.Op.OrderedFloat

Operators over float elements, including comparison operators

Forward Automatic Differentiation (FAD)

module F : functor (T : OpS) -> sig ... end

Construct a FAD-like module from a module of operators

module OrderedF : functor (T : OrderedOpS) -> sig ... end

Construct a FAD-like module from a module of operators over ordered elements

Backward Automatic Differentiation (BAD)

module B : functor (T : OpS) -> sig ... end

Construct a BAD-like module from a module of operators

module OrderedB : functor (T : OrderedOpS) -> sig ... end

Construct a BAD-like module from a module of operators over ordered elements

Automatic Taylor Expansion (TAD)

module T : functor (T : OpS) -> sig ... end

Construct a TAD-like module from a module of operators

Low level implementation

module Fadiff : sig ... end

module Badiff : sig ... end

Backward Automatic Differentiation (BAD)

module Tadiff : sig ... end

Automatic Taylor Expansion